SiteService inherits from BaseObjectService for event hooks
