+++ include/requirements.js
@@ -80,12 +80,14 @@ module.exports = function PB(config) {
     pb.SessionHandler = SessionModule(pb);
     pb.session        = new pb.SessionHandler(pb.SessionHandler.getSessionStoreInstance());
 
+    pb.BaseObjectService = require(path.join(config.docRoot, '/include/service/base_object_service.js'))(pb);
+
     //setup site service
     pb.SiteService = require(path.join(config.docRoot, '/include/service/entities/site_service.js'))(pb);
-    pb.sites = new pb.SiteService();
     pb.SiteQueryService = require(path.join(config.docRoot, '/include/service/entities/site_query_service.js'))(pb);
 
-    //setup object services
+
+  //setup object services
     pb.SimpleLayeredService         = require(path.join(config.docRoot, '/include/service/simple_layered_service.js'))(pb);
     pb.MemoryEntityService          = require(path.join(config.docRoot, '/include/service/memory_entity_service.js'))(pb);
     pb.CacheEntityService           = require(path.join(config.docRoot, '/include/service/cache_entity_service.js'))(pb);
@@ -110,7 +112,6 @@ module.exports = function PB(config) {
     pb.TokenAuthentication            = Authentication.TokenAuthentication;
 
     //setup user service
-    pb.BaseObjectService = require(path.join(config.docRoot, '/include/service/base_object_service.js'))(pb);
     pb.UserService       = require(path.join(config.docRoot, '/include/service/entities/user_service.js'))(pb);
     Object.defineProperty(pb, 'users', {
         get: function() {
