+++ include/service/jobs/sites/site_create_edit_job.js
@@ -87,7 +87,7 @@ module.exports = function SiteCreateEditJobModule(pb) {
         var tasks     = [
             //set site to active in mongo
             function(callback) {
-                var sos = new pb.SiteObjectService();
+                var SiteService = new pb.SiteService();
                 var dao = new pb.DAO();
                 dao.loadByValue('uid', mySite.uid, 'site', function(err, site) {
                     if(util.isError(err)) {
@@ -101,7 +101,7 @@ module.exports = function SiteCreateEditJobModule(pb) {
                     site.hostname = mySite.hostname || site.hostname;
                     site.displayName = mySite.displayName || site.displayName;
 
-                    sos.save(site, function(err, result) {
+                    SiteService.save(site, function(err, result) {
                         if(util.isError(err)) {
                             return cb(err, null);
                         }
